Create  utilizing babel
This code defines a component that has a text area where you can enter your code, and a button to compile the code. When the button is clicked, the handleCompile function is called which uses the Babel library to compile the code to a supported version of JavaScript. The compiled code is then displayed in a read-only text area